Combine all block files into index.js
Updates the carousel block to follow the same coding standards as other
blocks in the theme. Also adds some editor styling and re-adds margin
below the carousel element which was lost earlier.
